requirements.txt
aiohttp==3.9.1
flask==3.0.0
requests==2.31.0
selenium==4.16.0
webdriver-manager==4.0.1
torch
transformers==4.36.2
python-dotenv==1.0.0
asyncio==3.4.3
aiosignal==1.3.1
Werkzeug==3.0.1
Jinja2==3.1.2
click==8.1.7
charset-normalizer==3.3.2
certifi==2023.11.17
idna==3.6
urllib3==2.1.0
attrs==23.1.0
frozenlist==1.4.1
multidict==6.0.4
yarl==1.9.4
cloudinary
gunicorn